Front and Backend completion
UPDATE: frontend contents, 1/4 of 2000+ data updated, fixed unnecessary
CRUD bugs.

TO BE FIXED: previous direction route not removed, the exist data
notification and functionality, and searching on the map(speed,
stability and content).

TO BE ADDED: resident head members(if possible), data logs to be added
and updated(included members if possible) and frontend polishes(esp.
UI).